+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- dep: bump configuration to 2.0.0 and resolve type issues
+
 ### 2.1.1
 
 - fix: getProof error, MessageStatus enum